Fixed indentation, semi colons, socket server turned into class and adjusted to api doc
In addition create room no longer needs a host player and I added remove player functionality to only associate 1 player with each socket. This remove player call would be when a user joins a room creates a player and then hits the back button to edit so we need to remove their character from the player list without disconnecting.